updated receiver.cpp

+++ receiver.cpp
@@ -52,7 +52,7 @@ int main(int argc, char *argv[]) {
   socklen_t len = sizeof(client_addr);
 
   while(1) {
-    int n = recvfrom(sock, buf, sizeof(buf), 0, (sockaddr*)&client_addr, &len);
+    ssize_t n = recvfrom(sock, buf, sizeof(buf), 0, (sockaddr*)&client_addr, &len);
     if (n<0) {
       cerr << "[receiver] recv error: " << strerror(errno) << "\n";
       continue;
